+++ stock_move_invoice/__manifest__.py
@@ -3,7 +3,7 @@
 #
 #    Cybrosys Technologies Pvt. Ltd.
 #
-#    Copyright (C) 2021-TODAY Cybrosys Technologies(<https://www.cybrosys.com>).
+#    Copyright (C) 2020-TODAY Cybrosys Technologies(<https://www.cybrosys.com>).
 #    Author: Sayooj A O(<https://www.cybrosys.com>)
 #
 #    You can modify it under the terms of the GNU AFFERO
@@ -43,5 +43,5 @@
     'images': ['static/description/banner.png'],
     'installable': True,
     'application': True,
-    'sequence':1
+    'sequence': 1
 }