+++ stock_move_invoice/models/res_config_settings.py
@@ -23,8 +23,17 @@ from odoo import fields, models
 
 
 class Settings(models.TransientModel):
+    """Inheriting model res.config.settings to add journal fields"""
     _inherit = 'res.config.settings'
-    customer_journal_id = fields.Many2one('account.journal', string='Customer Journal',
-                                          config_parameter='stock_move_invoice.customer_journal_id')
-    vendor_journal_id = fields.Many2one('account.journal', string='Vendor Journal',
-                                        config_parameter='stock_move_invoice.vendor_journal_id')
+
+    customer_journal_id = fields.Many2one('account.journal',
+                                          string='Customer Journal',
+                                          config_parameter=
+                                          'stock_move_invoice.'
+                                          'customer_journal_id',
+                                          help='To add customer journal')
+    vendor_journal_id = fields.Many2one('account.journal',
+                                        string='Vendor Journal',
+                                        config_parameter=
+                                        'stock_move_invoice.vendor_journal_id',
+                                        help='To add vendor journal')