+++ stock_move_invoice/doc/RELEASE_NOTES.md
@@ -3,4 +3,9 @@
 #### 06.11.2020
 #### Version 16.0.1.0.0
 ##### ADD
-- Initial commit
+- Initial commit for Invoice From Stock Picking
+
+#### 19.07.2023
+#### Version 16.0.1.0.0
+##### UPDATE
+- Update for Invoice From Stock Picking